renamed packages, removed _core tags, added beta tags
