+++ jobs/v4/pom.xml
@@ -40,7 +40,6 @@
     </dependencyManagement>
 
     <dependencies>
-        <!-- [START dependencies] -->
         <dependency>
             <groupId>com.google.apis</groupId>
             <artifactId>google-api-services-jobs</artifactId>
@@ -64,7 +63,6 @@
             <artifactId>junit</artifactId>
             <version>4.13-rc-2</version>
         </dependency>
-
 
     </dependencies>
 </project>
